% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addArms.R
\name{addArms}
\alias{addArms}
\title{Add Arms to Trial}
\usage{
addArms(trial, sample_ratio, ...)
}
\arguments{
\item{trial}{an object returned from \code{trial()}.}

\item{sample_ratio}{integer vector. Sample ratio of the newly added arm(s)
for permuted block randomization. It will be appended to existing sample
ratio in the trial.}

\item{...}{one or more arm objects returned from \code{arm()}. One exception
in \code{...} is an argument \code{enforce}. When \code{enforce = TRUE},
sample ratio of newly added arm. It rolls back all patients after the time
when this function is called, i.e. redo randomization for unenrolled
patients. This can be useful to add arms one by one when creating a trial.
Note that we can run \code{addArms(trial, sample_ratio1, arm1)} followed
by \code{addArms(trial, sample_ratio2, enforce = TRUE, arm2)}.
We would expected similar result with
\code{addArms(trial, c(sample_ratio1, sample_ratio2), arm1, arm2)}. Note
that these two method won't return exactly the same trial because
randomization_queue were generated twice in the first approach but only
once in the second approach. But statistically, they are equivalent and
of the same distribution.}
}
\value{
\code{NULL}
}
\description{
Add one or more arms to a trial. Patients enrolled after the time when this
function is called will also be randomized into newly added arm(s).
This is a wrapper function of public
method \code{trial$add_arms(sample_ratio, ...)}.
See \code{?TrialSimulator::Trials} if you are comfortable to work with
\code{R6} class.
}
\examples{
## see an upcoming vignette of dose ranging trial
## briefly speaking, call this function
## 1) after a trial is defined/initialized, i.e. after calling trial()
##    to register arms to a new trial, or
## 2) within an action function to add new arm(s) to a running trial
##    dynamically and adaptively

## Example 1: add arms to newly created trial
ep <- endpoint(name = 'pfs', type = 'tte', generator = rexp, rate = log(2)/10)
pbo <- arm(name = 'placebo')
addEndpoints(pbo, ep)

ep <- endpoint(name = 'pfs', type = 'tte', generator = rexp, rate = log(2)/13)
trt <- arm(name = 'treated')
addEndpoints(trt, ep)

rates <- data.frame(
  end_time = c(10, Inf),
  piecewise_rate = c(50, 80))

trial <- trial(
  name = '3.1415',
  n_patients = 100,
  duration = 20,
  enroller = StaggeredRecruiter,
  accrual_rate = rates
)

addArms(trial, sample_ratio = c(1, 2), trt, pbo) # trt:pbo = 1:2

trial

}
