% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEndpoints.R
\name{addEndpoints}
\alias{addEndpoints}
\title{Adding Endpoints to an Arm}
\usage{
addEndpoints(arm, ...)
}
\arguments{
\item{arm}{an object returned from \code{arm()}.}

\item{...}{one or more objects returned from \code{endpoint()}.}
}
\value{
\code{NULL}
}
\description{
Add one or more endpoints to an arm. This is a wrapper function of public
method \code{arm$add_endpoints(...)}. See \code{?TrialSimulator::Arms} if
you are comfortable to work with \code{R6} class.
}
\examples{
## see vignette:
## https://zhangh12.github.io/TrialSimulator/articles/defineArms.html
## replace arm$add_endpoints(...) with addEndpoints(arm, ...) in that vignette.

}
