% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMilestones.R
\name{addMilestones}
\alias{addMilestones}
\title{Registering Milestones with Listener}
\usage{
addMilestones(listener, ...)
}
\arguments{
\item{listener}{an object returned from \code{listener()}.}

\item{...}{milestone objects. Order in \code{...} matters
as they are scanned and triggered in that order. It is users' responsibility
to use reasonable order when calling this function, otherwise,
the result can be problematic.}
}
\value{
\code{NULL}
}
\description{
Assigning one or multiple milestones to a listener so that it can trigger
them during a trial.
}
