% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arm.R
\name{arm}
\alias{arm}
\title{Defining an Arm}
\usage{
arm(name, ...)
}
\arguments{
\item{name}{name of arm, which is the arm's label in generated data}

\item{...}{subset condition that is compatible with \code{dplyr::filter}.
This can be used to specify inclusion criteria of an arm.
By default it is not specified, i.e. all data generated by the generator
will be used as trial data. More than one conditions can be
specified in \code{...}.}
}
\value{
a summary report (html) of the endpoints in the arm accounting for
subset conditions if executed in console
}
\description{
Define an arm in trial
}
\examples{
## see vignette:
## https://zhangh12.github.io/TrialSimulator/articles/defineArms.html

}
