% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source-edit.R
\name{build_chunk_map}
\alias{build_chunk_map}
\title{Build the chunk map}
\usage{
build_chunk_map(path, label)
}
\arguments{
\item{path}{`character` path to the source file}

\item{label}{`character` comment label of code chunk of interest}
}
\value{
line number for chunk "label"
}
\description{
Build a map between comment labels and their line number locations.
}
