% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendarTime.R
\name{calendarTime}
\alias{calendarTime}
\title{Triggering a Milestone by Calendar Time}
\usage{
calendarTime(time)
}
\arguments{
\item{time}{numeric. Calendar time to trigger a milestone of a trial.}
}
\value{
an object of R6 class \code{Condition}
}
\description{
Define the condition to trigger trial milestone by calendar time.
The milestone will be triggered when a trial has been running for at least
the specified duration. It can be used combined with conditions specified by
\link[TSLite]{enrollment} and \link[TSLite]{eventNumber}. Refer to vignette
for examples.

\url{https://zhangh12.github.io/TrialSimulator/articles/conditionSystem.html}
}
\examples{
## see vignette
## https://zhangh12.github.io/TrialSimulator/articles/conditionSystem.html
}
