% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrelatedPfsAndOs3.R
\name{CorrelatedPfsAndOs3}
\alias{CorrelatedPfsAndOs3}
\title{Generate Correlated PFS and OS Using the Three-States Model}
\usage{
CorrelatedPfsAndOs3(n, h01, h02, h12, pfs_name = "pfs", os_name = "os")
}
\arguments{
\item{n}{integer. Number of observations.}

\item{h01}{constant transition hazard from state "initial" to state "progression".}

\item{h02}{constant transition hazard from state "initial" to state "death".}

\item{h12}{constant transition hazard from state "progression" to state "death".}

\item{pfs_name}{column name of PFS in returned data frame. It must be
consistent with the \code{name} argument in the function \code{endpoint()}.}

\item{os_name}{column name of OS in returned data frame. It must be
consistent with the \code{name} argument in the function \code{endpoint()}.}
}
\value{
A data frame of four columns, including PFS, OS and their event indicators.
The event indicators are all 1s.
}
\description{
Generate Correlated PFS and OS Using the Three-States Model
}
\examples{

## See examples in ?endpoint

}
