% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source-edit.R
\name{edit}
\alias{edit}
\title{Jump to edit source}
\usage{
edit(label, script)
}
\arguments{
\item{label}{`character` comment label of code chunk of interest}

\item{script}{`character` the file to be opened}
}
\description{
Open the source code at appropriate location to edit in RStudio.
}
