% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source-fill.R
\name{fill}
\alias{fill}
\title{Fill symbol values}
\usage{
fill(
  path,
  scenario,
  out_path = sub("\\\\.R$", "_filled.R", path),
  keep_roxygen_comments = FALSE
)
}
\arguments{
\item{path}{`character` path to R script containing `simulate()` function source}

\item{scenario}{`tibble` one scenario per row tibble}

\item{out_path}{`character` path for temporary output files}

\item{keep_roxygen_comments}{`logical` whether to keep roxygen comments}
}
\description{
Fill symbol values from the scenario tibble. `simulate()` function from
each template has labeled chunks. To run this chunk in .qmd files one needs
values. These values are supplied from scenario tibble. Ensure that all
chunks mentioned in `simulate()` function are used, and chunk used in .qmd
file exist in `simulate()` funciton.
}
