% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLockedData.R
\name{getLockedData}
\alias{getLockedData}
\title{Remove Arms from a Trial}
\usage{
getLockedData(trial, milestone_name)
}
\arguments{
\item{trial}{a trial object returned from \code{trial()}.}

\item{milestone_name}{character. Milestone name of which the locked
data to be extracted.}
}
\value{
\code{NULL}
}
\description{
Remove Arms from a Trial
}
