% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOutput.R
\name{getOutput}
\alias{getOutput}
\title{Retrieving Output of Simulations}
\usage{
getOutput(controller, cols = NULL, simplify = TRUE)
}
\arguments{
\item{controller}{a controller object returned from \code{controller()}.}

\item{cols}{vector of character. Columns to be returned. If \code{NULL},
all existing columns are returned.}

\item{simplify}{logical. Return value rather than a data frame of one
column when \code{length(cols) == 1} and \code{simplify == TRUE}.}
}
\value{
a data frame or a scalar depending on length of \code{cols} and
\code{simplify}.
}
\description{
Returning a data frame of all current outputs saved by calling \code{save},
or intermediate results saved by \code{TSLite} automatically, including
triggering time of milestones, and number of observations/events of endpoints
at milestones. Note that this data frame is updated dynamically during a
trial, so it can return different values depending on when you call it.
}
