% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listener.R
\name{listener}
\alias{listener}
\title{Defining a Listener}
\usage{
listener(silent = FALSE)
}
\arguments{
\item{silent}{logical. \code{TRUE} to mute messages.}
}
\value{
a listener object
}
\description{
\code{TSLite} relies on a listener to monitor a running trial. It determine
the earliest time that can trigger a milestone, and execute user-defined
action. Milestones need to be assigned to a listener before it can do the
job.
}
\examples{
listener <- listener()
}
