% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milestone.R
\name{milestone}
\alias{milestone}
\title{Defining Milestones of Trial}
\usage{
milestone(name, when, action = doNothing)
}
\arguments{
\item{name}{character. Name of milestone.}

\item{when}{condition to check if this milestone should be
triggered.}

\item{action}{function to execute when the milestone is triggered.
If no action to be executed but simply need to record triggering time and
number of observations/events at the triggering time, \code{action} can be
its default value, a built-in function \code{doNothing}.}
}
\value{
a milestone object
}
\description{
Defining a milestone of a trial. This includes the condition to trigger it
and the action to be taken automatically. Refer to vignettes for more details.

\url{https://zhangh12.github.io/TrialSimulator/articles/conditionSystem.html}

\url{https://zhangh12.github.io/TrialSimulator/articles/actionFunctions.html}
}
\examples{
## See the two vignettes in description.

}
