% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeArms.R
\name{removeArms}
\alias{removeArms}
\title{Removing Arms from a Trial}
\usage{
removeArms(trial, arms_name)
}
\arguments{
\item{trial}{a trial object returned from \code{trial()}.}

\item{arms_name}{character vector. Name of arms to be removed from a trial}
}
\value{
\code{NULL}
}
\description{
remove arms from a trial. Planned but unenrolled patients will be
randomized to remaining arms. This function may be used with futility
analysis, dose selection, enrichment design (sub-population) or
interim analysis (early stop for efficacy), etc.
}
