% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset.R
\name{reset}
\alias{reset}
\title{Reset a Controller}
\usage{
reset(controller)
}
\arguments{
\item{controller}{a controller object returned from \code{controller()}.}
}
\value{
\code{NULL}
}
\description{
Reset the trial and listener registered with the controller before calling
\code{run(controller, n = ...)} again.
}
\examples{

## run(controller, n =1)
## reset(controller)
## run(controller, n = 10)
}
