% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Running a Trial}
\usage{
run(controller, n = 1, plot_event = TRUE, silent = FALSE)
}
\arguments{
\item{controller}{a controller object returned from \code{controller()}.}

\item{n}{integer. Number of replicates of simulation. \code{n = 1} by default.}

\item{plot_event}{logical. Create event plot. This should be set to
\code{FALSE} when \code{n} is greater than 1 as it is time-consuming and
is unnecessary.}

\item{silent}{logical. \code{TRUE} if muting all messages during a
trial. Note that warning messages are still displayed.}
}
\value{
\code{NULL}
}
\description{
Running a trial for once or multiple times.
}
