% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setDuration.R
\name{setDuration}
\alias{setDuration}
\title{Set Duration of a Trial}
\usage{
setDuration(trial, duration)
}
\arguments{
\item{trial}{a trial object returned from \code{trial()}.}

\item{duration}{new duration of a trial. It must be longer than the current
duration.}
}
\value{
\code{NULL}
}
\description{
set trial duration in an adaptive designed trial. All patients enrolled
before resetting the duration are truncated (for non-tte endpoints) or
censored (for tte endpoints) at the original duration. Remaining patients
are re-randomized. New duration must be longer than the old one. This
function is usually called in an action function.
}
