% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StaggeredRecruiter.R
\name{StaggeredRecruiter}
\alias{StaggeredRecruiter}
\title{Generate Enrollment Time from Piecewise Constant Uniform Distribution}
\usage{
StaggeredRecruiter(n, accrual_rate)
}
\arguments{
\item{n}{number of patients}

\item{accrual_rate}{a data frame of columns
\code{end_time}:
    End time for a constant rate in a time window. The start time of
    the first time window is 0.
\code{piecewise_rate}:
    A constant rate in a time window. So the number of patients
    being recruited in that window is window length x \code{piecewise_rate}}
}
\value{
a vector of enrollment time
}
\description{
determine enrollment time for patients with piecewise constant rates in
time windows.
}
\examples{
accrual_rate <- data.frame(
  end_time = c(12, 13:17, Inf),
  piecewise_rate = c(15, 15 + 6 * (1:5), 45)
)

accrual_rate <- data.frame(
  end_time = c(3, 4, 5, 8, Inf),
  piecewise_rate = c(1, 2, 2, 3, 4)
)

StaggeredRecruiter(30, accrual_rate)

}
