% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial.R
\name{trial}
\alias{trial}
\title{Defining a Trial}
\usage{
trial(
  name,
  n_patients,
  duration,
  seed = NULL,
  enroller,
  dropout = NULL,
  silent = FALSE,
  ...
)
}
\arguments{
\item{name}{character. Name of trial.}

\item{n_patients}{integer. Planned number of patients could be enrolled
to the trial. It can be updated during a trial through adaptation functions.}

\item{duration}{Numeric. Planned trial duration. It can be updated during a
trial through adaptation functions. Note that this argument does not assume
an unit, make sure it is aligned with other parameters in trial/endpoint/arm
settings.}

\item{seed}{random seed. If \code{NULL}, \code{TSLite} will set seed for
each simulated trial automatically. Recommend always setting it to
\code{NULL} unless when debugging.}

\item{enroller}{a function returning a vector of patient enrollment time.
Its first argument is the number of enrolled patients. Recommend using
the built-in enroller \code{StaggeredRecruiter}.}

\item{dropout}{a function returning a vector of dropout time for
patients. Its first argument is the number of enrolled patients.}

\item{silent}{logical. \code{TRUE} to mute messages.}

\item{...}{arguments of \code{enroller} and \code{dropout}.}
}
\value{
a trial object
}
\description{
Defining a trial with planned settings with some of those can be adaptively
altered later.
}
\examples{
trial <- trial(
  name = '3.1415',
  n_patients = 650,
  duration = 24,
  seed = NULL, # recommended
  enroller = StaggeredRecruiter, # recommended
  dropout = rweibull,
  accrual_rate =
    data.frame(
      end_time = c(10, Inf),
      piecewise_rate = c(30, 40)), # argument of enroller
  shape = 2, # argument of dropout
  scale = 40, # argument of dropout
)

trial

}
