% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSLite-package.R
\docType{package}
\name{TSLite-package}
\alias{TSLite}
\alias{TSLite-package}
\title{TSLite: Infrastructure of Simulation in Planning Clinical Trials}
\description{
TSLite provides a lightweight extension to the TrialSimulator package. While TrialSimulator has already undergone substantial testing and offers a comprehensive infrastructure with advanced tools for clinical trial simulation, TSLite focues on delivering a thoroughly tested subset of its functionality. It wraps validated functions to support reliable and efficient trial simulation. Over time, more tools and functions from TrialSimulator will be incorporated into TSLite once they have passed rigorous testing.
}
\author{
\strong{Maintainer}: Han Zhang \email{han.zhang@boehringer-ingelheim.com}

Other contributors:
\itemize{
  \item Saumil Shah \email{saumil.shah@boehringer-ingelheim.com} [contributor]
}

}
\keyword{internal}
